Revert "Full rewrite"
