--- .gitattributes
@@ -1,2 +1,0 @@
-# LF Normalization
-* text=CLRF
